refactor(admin): split live ops panel modules
Move live ops rendering out of the oversized panel file into primitives, typed initial data, health panels, personal data panels, and the process panel while preserving the existing live-ops-panels barrel import used by the page.

Verified with pnpm exec prettier --check apps/admin/src/app/ops/live-ops-panels.tsx apps/admin/src/app/ops/live-ops-primitives.tsx apps/admin/src/app/ops/live-ops-health-panels.tsx apps/admin/src/app/ops/live-ops-personal-panels.tsx apps/admin/src/app/ops/live-ops-process-panel.tsx apps/admin/src/app/ops/live-ops-system-panels.tsx apps/admin/src/app/ops/live-ops-types.ts; pnpm --filter @anipotts/admin typecheck; pnpm --filter @anipotts/admin lint; pnpm --filter @anipotts/admin build; git diff --check -- apps/admin/src/app/ops/live-ops-panels.tsx apps/admin/src/app/ops/live-ops-primitives.tsx apps/admin/src/app/ops/live-ops-health-panels.tsx apps/admin/src/app/ops/live-ops-personal-panels.tsx apps/admin/src/app/ops/live-ops-process-panel.tsx apps/admin/src/app/ops/live-ops-system-panels.tsx apps/admin/src/app/ops/live-ops-types.ts; pnpm validate; pnpm dlx knip --no-exit-code.
