+++ apps/admin/src/app/ops/live-ops-system-panels.tsx
@@ -1,0 +1,11 @@
+export {
+  AgentsSection,
+  RudySection,
+  SyncthingSection,
+  VitalsSection,
+} from "./live-ops-health-panels";
+export {
+  PresenceSection,
+  SessionsSection,
+  VaultSection,
+} from "./live-ops-personal-panels";
